Integrate CH32V003/V203 memory analysis and comprehensive squeeze testing
Major Updates:
- ✅ CH32V203: Full Embassy support, 21 HAL tests passing
- 🟡 CH32V003: Embassy viable at 95% RAM usage, 8-element queue
- 🧪 Squeeze tests integrated into HAL (15→21 tests)
- 📊 Real memory footprint measurements vs theoretical calculations
- 🔧 RISC-V critical section implementation
- ⚙️ Generic queue size implementation (8/64 elements)

Key Learning:
RISC-V runtime auto-allocates remaining RAM to stack, making
actual build testing essential over theoretical calculations.

Technical Changes:
- Remove obsolete tests/ directory, integrate into hal_tests.rs
- Fix portable-atomic configuration for RISC-V targets
- Update Cargo workspace structure (3 members)
- Embassy task arena reduced from 8KB→1KB
- Memory-optimized configurations for both targets

Test Coverage:
- 6 new squeeze test cases covering all keyer modes
- PaddleInput integration with FSM verification
- Release pattern and timing boundary testing
- All 21 tests passing on both target architectures

+++ keyer-core/src/test_utils.rs
@@ -189,7 +189,7 @@ pub mod paddle_simulator {
     pub fn execute_pattern(paddle: &PaddleInput, pattern: &PaddlePattern, debounce_ms: u64) {
         for event in &pattern.events {
             // In real test, this would use virtual time
-            paddle.update(event.side, event.pressed, debounce_ms);
+            paddle.update(event.side, event.pressed, debounce_ms as u32);
         }
     }
 }
